requirements.txt
python-pptx>=1.0
PyYAML>=6.0
streamlit>=1.40
